Android support in gradle plugin (fix issue #266)
Previously the `java` plugin was automatically applied when
configuring the task dependencies. This was not compatible with
the `android` plugin. Now this plugin will inspect the build
to determine the right configuration to target.

Added the example provided by @samskiter that demonstrated the
bug and now shows a working sample.