Add Gradle test example & fix source file directory handling
Added an example for testing the Gradle plugin by executing an external
build against a known project. This will allow adding more scenarios
based on bug reports to verify configurations. Unfortunately the more
mocking approach, via ProjectBuilder, does not seem to work correctly
in the Maven bootstrap.

It appears that the type change of `File` to `URL` in GenerationConfig
`source` broke the plugin. It calls File's `mkdir` if necessary during
configuration. As URL lacks that method, the plugin breaks unless the
`source` is explicitly configured. This is now resolved.

Note that due to the above mentioned bug, the CI may break as the
SNAPSHOT dependency won't be in the local repository, so the plugin
test will break by using an old plugin version. An `mvn install -DskipTests`
followed by a build shows the test passing.