Fix edge case of getter/setter naming when second char is uppercase
A little known part of the been spec, since a getter/setter with a
capital letter in second position indicates that the first char is also
a capital. If the first char is in fact not a capital, then it needs to
be lowercase in the getter/setter also.

See also:

* https://stackoverflow.com/questions/8969112/confused-about-naming-of-javabean-properties-with-respect-to-getters-and-setter
* https://github.com/FasterXML/jackson-databind/issues/653
* https://bugs.eclipse.org/bugs/show_bug.cgi?id=322223

Fixes #756