Remove runtime dependency on Moshi
Moshi is now supported without requiring Moshi types on the classpath,
so there is no longer a direct dependency on Moshi inside the plugin.

As part of this change, Annotator method signatures have had to change
to ensure that current class/enum is always passed to the annotator
method. This allows the annotator to always create a direct class ref,
because it will always have access to the owning code model.

Closes #874
Closes #876