Separate javaType into 'existingJavaType' and 'javaType'
This is a pretty major breaking change. If you use 'javaType' to refer
to a primitive type or some existing class that you don't want
generated, you now need to use 'existingJavaType' instead. The
'javaType' rule is now used to give an explicit name to a new class that
will be generated.

Closes #828