Clear schema cache between JSON files
When using JSON or YAML files (not schemas, but example data) the
schemas will not have a proper file path in their ID, as there is no
actual schema file. All schema IDs will be fragments only (relative
references within the document). Once generation is complete for a file,
these schemas should be thrown away and not reused.

Closes #1427