Make equals and hashCode generation deterministic (order of expressions)
Order of lines of code in equals and hashCode was non-deterministic.
This means that the generated files with code have changed between generations (even if nothing has changed in the schema itself).

After the change, the order of these lines will be the same as the order of fields in the schema.