A demonstration of how we could transform allOf type keywords without needing to worry about cycles in the graph.
This is initial demo is no were near complete, but does should the idea of repeatedly applying a transform to the schema graph to scatter
properties until no more changes result from the transform process.  This algorithm is parallelizable in both the gather and scatter steps.

To install this, we would need to place it between the rules and the SchemaStore.  A function for loading documents would also need to be added that could load documents after a step where the document was needed but not already loaded.